nuplayer: set maxBuffers to 1 in byte stream mode
For MP3 and AAC clips, byte stream mode is enabled
by default in which the parser gives a huge chunk of
data per read. In this case, reading 64 buffers is
unnecessary as it leads to delay in data being written
to HAL which causes a gap during gapless transition.

Set maxBuffers to 1 when byte stream mode is enabled.

Change-Id: I4f51bbc4c53d17f62cab571b83ca015cea430fa4